% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazettes.R
\name{gazettes}
\alias{gazettes}
\title{gazettes
Descrição da função
parâmetros}
\usage{
gazettes(
  since = NULL,
  until = NULL,
  keywords = NULL,
  territory_id = NULL,
  offset = 0,
  size = 10
)
}
\description{
gazettes
Descrição da função
parâmetros
}
\details{
retur descrição do resultado
export
}
