% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazettes.R
\name{get_gazettes}
\alias{get_gazettes}
\title{get_gazettes}
\usage{
get_gazettes(
  since = NULL,
  until = NULL,
  keywords = NULL,
  territory_id = NULL,
  offset = 0,
  size = 10
)
}
\arguments{
\item{since}{data a partir da qual se deseja pesquisar, em formato YYY-MM-DD}

\item{until}{data ate quando, em formato YYY-MM-DD}

\item{keywords}{palavras-chave que serão pesquisadas nos diarios oficiais}

\item{territory_id}{número dos municipios, seguindo os codigos da tabela}

\item{offset}{sempre igual a 0}
}
\value{
um arquivo json contendo os diários oficiais para os parâmetros indicados
}
\description{
Requisita diarios oficiais a API do Querido Diário e retorna um json
}
