% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{territories}
\alias{territories}
\title{territories}
\format{
um data frame com 12 e 3 variaveis:

\describe{\item{territory_id}{codigo ID.}
\item{territory_name}{nome da respectiva capital}
\item{territory_state}{nome do estado respectivo.}
}
}
\source{
territories
}
\usage{
territories
}
\description{
Um conjunto de dados que contem informaçoes sobre capitais e os respectivos ids.
Na base não consta todos os estados brasileiros.
}
\keyword{datasets}
