% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazettes.R
\name{get_gazettes}
\alias{get_gazettes}
\title{get_gazettes}
\usage{
get_gazettes(
  since = NULL,
  until = NULL,
  keywords = NULL,
  territory_id = NULL,
  offset = 0,
  size = NULL
)
}
\arguments{
\item{since}{data a partir da qual se deseja pesquisar, em formato YYY-MM-DD}

\item{until}{data ate quando, em formato YYY-MM-DD}

\item{keywords}{palavras-chave que serão pesquisadas nos diarios oficiais}

\item{territory_id}{numero dos municipios, seguindo os codigos da tabela}

\item{offset}{sempre igual a 0}

\item{size}{retorna uma quantidade de itens}
}
\value{
um arquivo json contendo os diarios oficiais para os parametros indicados
}
\description{
Requisita diarios oficiais a API do Querido Diario e retorna um json
}
\seealso{
data frame com IDs dos \link{territories}
}
