% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gazettes.R
\name{get_gazettes}
\alias{get_gazettes}
\title{get_gazettes}
\usage{
get_gazettes(
  since = NULL,
  until = NULL,
  keywords = NULL,
  territory_id = NULL,
  offset = 0,
  size = 10
)
}
\arguments{
<<<<<<< HEAD
\item{since}{data a partir da qual, em formato YYY-MM-DD}

\item{until}{data ate quando, em formato YYY-MM-DD}

\item{keywords}{palavras-chave que sera pesquisadas nos diarios oficiais}

\item{territory_id}{numero dos municipios, seguindo os codigos da tabela do ibge}

\item{offset}{sempre igual a 0}

\item{size}{numero de a diarios oficiais desejados}
}
\value{
um arquivo json contendo os diarios oficiais para os parametros indicados
export
}
\description{
Requisita diarios oficiais para a API do Querido Diario e retorna um json
=======
\item{since}{data a partir da qual se deseja pesquisar, em formato YYY-MM-DD}

\item{until}{data até quando, em formato YYY-MM-DD}

\item{keywords}{palavras-chave que serão pesquisadas nos diários oficiais}

\item{territory_id}{número dos municípios, seguindo os códigos da tabela}

\item{offset}{sempre igual a 0}

\item{size}{numero de a diários oficiais desejados}
}
\value{
um arquivo json contendo os diários oficiais para os parâmetros indicados
export
}
\description{
Requisita diários oficiais à API do Querido Diário e retorna um json
>>>>>>> bd581248e42131603981f2df6900ca85e79a7c98
}
